CARF-498: Moved RCASP files into registration packages
