+++ conf/management.routes
@@ -1,3 +1,3 @@
 # management microservice specific routes
 
-GET        /view-rcasp/:carfId/:rcaspId          uk.gov.hmrc.carfmanagement.controllers.RcaspController.viewRcasp(carfId: String, rcaspId: String)
+GET        /view-rcasp/:carfId/:rcaspId          uk.gov.hmrc.carfregistration.controllers.RcaspController.viewRcasp(carfId: String, rcaspId: String)
